refactor: rename group member requests module
