+++ src/groups/group-member-requests/group-member-requests.resolver.ts
@@ -14,21 +14,21 @@ import { CurrentUser } from "../../auth/decorators/current-user.decorator";
 import { Dataloaders } from "../../dataloader/dataloader.types";
 import { User } from "../../users/models/user.model";
 import { Group } from "../models/group.model";
-import { MemberRequestsService } from "./member-requests.service";
-import { ApproveMemberRequestPayload } from "./models/approve-member-request.payload";
-import { CreateMemberRequestPayload } from "./models/create-member-request.payload";
-import { MemberRequest } from "./models/member-request.model";
+import { GroupMemberRequestsService } from "./group-member-requests.service";
+import { ApproveMemberRequestPayload } from "./models/approve-group-member-request.payload";
+import { CreateMemberRequestPayload } from "./models/create-group-member-request.payload";
+import { MemberRequest } from "./models/group-member-request.model";
 
 @Resolver(() => MemberRequest)
-export class MemberRequestsResolver {
-  constructor(private memberRequestsService: MemberRequestsService) {}
+export class GroupMemberRequestsResolver {
+  constructor(private groupMemberRequestsService: GroupMemberRequestsService) {}
 
   @Query(() => MemberRequest, { nullable: true })
   async memberRequest(
     @Args("groupId", { type: () => Int }) groupId: number,
     @CurrentUser() { id }: User
   ) {
-    return this.memberRequestsService.getMemberRequest({
+    return this.groupMemberRequestsService.getMemberRequest({
       user: { id },
       groupId,
     });
@@ -55,21 +55,21 @@ export class MemberRequestsResolver {
     @Args("groupId", { type: () => Int }) groupId: number,
     @CurrentUser() { id }: User
   ) {
-    return this.memberRequestsService.createMemberRequest(groupId, id);
+    return this.groupMemberRequestsService.createMemberRequest(groupId, id);
   }
 
   @Mutation(() => ApproveMemberRequestPayload)
   async approveMemberRequest(@Args("id", { type: () => Int }) id: number) {
-    return this.memberRequestsService.approveMemberRequest(id);
+    return this.groupMemberRequestsService.approveMemberRequest(id);
   }
 
   @Mutation(() => Boolean)
   async cancelMemberRequest(@Args("id", { type: () => Int }) id: number) {
-    return this.memberRequestsService.cancelMemberRequest(id);
+    return this.groupMemberRequestsService.cancelMemberRequest(id);
   }
 
   @Mutation(() => Boolean)
   async denyMemberRequest(@Args("id", { type: () => Int }) id: number) {
-    return this.memberRequestsService.denyMemberRequest(id);
+    return this.groupMemberRequestsService.denyMemberRequest(id);
   }
 }
