+++ src/groups/models/group.model.ts
@@ -19,7 +19,7 @@ import { Proposal } from "../../proposals/models/proposal.model";
 import { User } from "../../users/models/user.model";
 import { GroupConfig } from "../group-configs/models/group-config.model";
 import { GroupRole } from "../group-roles/models/group-role.model";
-import { MemberRequest } from "../member-requests/models/member-request.model";
+import { MemberRequest } from "../group-member-requests/models/group-member-request.model";
 
 @Entity()
 @ObjectType()
